Move the layout language into its own module.
