+++ ui-effects.cabal
@@ -24,6 +24,7 @@ library
                      , GL.Setup
                      , GL.Shader
                      , UI.Drawing
+                     , UI.Layout
                      , UI.View
                      , UI.Window
   build-depends:       base >= 4.7 && < 5
